+++ server/web/hooks.go
@@ -2,13 +2,12 @@ package web
 
 import (
 	"encoding/json"
-	"mime"
-	"net/http"
-	"path/filepath"
-
 	"github.com/beego/beego/v2/core/logs"
 	"github.com/beego/beego/v2/server/web/context"
 	"github.com/beego/beego/v2/server/web/session"
+	"mime"
+	"net/http"
+	"path/filepath"
 )
 
 // register MIME type with content type