+++ server/web/router.go
@@ -69,20 +69,8 @@ var (
 		"UNLOCK":    true,
 	}
 	// these web.Controller's methods shouldn't reflect to AutoRouter
-	exceptMethod = []string{"Abort", "CheckXSRFCookie", "CustomAbort", "DelSession",
-		"DestroySession", "Finish", "GetBool", "GetControllerAndAction",
-		"GetFile", "GetFiles", "GetFloat", "GetInt", "GetInt16",
-		"GetInt32", "GetInt64", "GetInt8", "GetSecureCookie", "GetSession",
-		"GetString", "GetStrings", "GetUint16", "GetUint32", "GetUint64",
-		"GetUint8", "HandlerFunc", "Init", "Input",
-		"IsAjax", "Mapping", "ParseForm",
-		"Prepare", "Redirect", "Render", "RenderBytes",
-		"RenderString", "SaveToFile", "SaveToFileWithBuffer", "SaveToFileWithBuffer",
-		"ServeFormatted", "ServeJSON", "ServeJSONP", "ServeXML", "ServeYAML",
-		"SessionRegenerateID", "SetData", "SetSecureCookie", "SetSession", "StartSession",
-		"StopRun", "URLFor", "URLMapping", "XSRFFormHTML",
-		"XSRFToken",
-	}
+	// see registerControllerExceptMethods
+	exceptMethod = initExceptMethod()
 
 	urlPlaceholder = "{{placeholder}}"
 	// DefaultAccessLogFilter will skip the accesslog if return true
@@ -114,6 +102,17 @@ func (l *logFilter) Filter(ctx *beecontext.Context) bool {
 // ExceptMethodAppend to append a slice's value into "exceptMethod", for controller's methods shouldn't reflect to AutoRouter
 func ExceptMethodAppend(action string) {
 	exceptMethod = append(exceptMethod, action)
+}
+
+func initExceptMethod() []string {
+	res := make([]string, 0, 32)
+	c := &Controller{}
+	t := reflect.TypeOf(c)
+	for i := 0; i < t.NumMethod(); i++ {
+		m := t.Method(i)
+		res = append(res, m.Name)
+	}
+	return res
 }
 
 // ControllerInfo holds information about the controller.